pyproject.toml
[tool.poetry]
name = "automation-test"
version = "0.1.0"
description = ""
authors = ["michkail <michkail@sateknologi.id>"]
readme = "README.md"
packages = [{include = "automation_test"}]

[tool.poetry.dependencies]
python = "^3.10"
appium-python-client = "^2.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
